+++ stratosphere-addons/spargel/src/main/java/eu/stratosphere/pact/spargel/VertexUpdateFunction.java
@@ -12,7 +12,7 @@
  * specific language governing permissions and limitations under the License.
  *
  **********************************************************************************************************************/
-package eu.stratosphere.pact.vertexcentric;
+package eu.stratosphere.pact.spargel;
 
 import java.io.Serializable;
 import java.util.Iterator;
