+++ stratosphere-addons/spargel/src/main/java/eu/stratosphere/pact/spargel/examples/connectedcomponents/LongLongInputFormat.java
@@ -13,7 +13,7 @@
  *
  **********************************************************************************************************************/
 
-package eu.stratosphere.pact.vertexcentric.examples.cc;
+package eu.stratosphere.pact.spargel.examples.connectedcomponents;
 
 import eu.stratosphere.pact.common.io.TextInputFormat;
 import eu.stratosphere.pact.common.type.PactRecord;
